# be/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Private Clinic API
  version: 0.17.0
  description: Public contract for the private clinic management system.
servers:
  - url: http://localhost:5000
paths:
...
  /api/v1/patient-access/reference-data:
    get:
      operationId: getPatientLinkReferenceData
      summary: List active branches that can receive a patient-link request
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Branch choices for an authenticated patient.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientLinkReferenceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/patient-access:
    get:
      operationId: getPatientAccess
      summary: List accessible patient profiles, delegated users, and the caller's link requests
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current portal access and request history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientAccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/patient-access/requests:
    post:
      operationId: requestPatientLink
      summary: Request access to an existing patient profile
      description: A request that does not match a patient is accepted as a decoy and never appears in the staff queue.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientLinkRequest'
      responses:
        '202':
          description: The request was accepted without revealing whether the patient reference matched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientLinkRequestAcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/patient-access/requests/{requestId}:
    delete:
      operationId: cancelPatientLinkRequest
      summary: Cancel the authenticated patient's pending link request
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientLinkRequestId'
      responses:
        '200':
          description: The pending request was cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientLinkCancelledResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/patient-access/links/{linkId}:
    delete:
      operationId: revokeOwnPatientLink
      summary: Revoke the caller's non-SELF link or delegated access to the caller's SELF profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientLinkId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientLinkReasonRequest'
      responses:
        '200':
          description: The link was revoked and can no longer book appointments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientLinkRevokedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/admin/patients/reference-data:
    get:
      operationId: getPatientReferenceData
      summary: List branches where the actor may manage patient administrative records
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Manageable branches.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientReferenceResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/admin/patients:
    post:
      operationId: createPatient
      summary: Register a patient after a scoped duplicate check
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePatientRequest' }
      responses:
        '201':
          description: Patient registered.
          headers:
            ETag: { $ref: '#/components/headers/CatalogEtag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/patients/{patientId}:
    get:
      operationId: getPatient
      summary: Read an administrative record in its registration branch
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/BranchPublicIdQuery'
      responses:
        '200':
          description: Patient administrative details.
          headers:
            ETag: { $ref: '#/components/headers/CatalogEtag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientDetailResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updatePatient
      summary: Update administrative details with optimistic concurrency
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/BranchPublicIdQuery'
        - $ref: '#/components/parameters/CatalogIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientWriteRequest' }
      responses:
        '200':
          description: Patient updated.
          headers:
            ETag: { $ref: '#/components/headers/CatalogEtag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
...
  /api/v1/reception/patients:
    get:
      operationId: searchReceptionPatients
      summary: Search active patients for walk-in reception at an authorized branch
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BranchPublicIdQuery'
        - name: query
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 100 }
      responses:
        '200':
          description: Matching active patients.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReceptionPatientListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  parameters:
...
    PatientId:
      name: patientId
      in: path
      required: true
      schema: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
...
    CatalogIfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing the Base64 rowversion of a catalog resource.
      schema:
        type: string
        pattern: '^"[A-Za-z0-9+/]{11}="$'
...
    BranchPublicIdQuery:
      name: branchPublicId
      in: query
      required: true
      schema:
        type: string
        pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID that makes retries with the same operation payload return the original result.
      schema:
        type: string
        format: uuid
    PatientLinkRequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PatientLinkId:
      name: linkId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    CatalogEtag:
      description: Strong ETag containing the Base64 SQL rowversion.
      schema:
        type: string
        pattern: '^"[A-Za-z0-9+/]{11}="$'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication data is missing, invalid, expired, reused, or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: The account is unavailable for login.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: The requested resource does not exist in the caller's scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: The operation conflicts with uniqueness, concurrency, or account safety rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    PreconditionRequired:
      description: The If-Match precondition is required for this update.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    TooManyRequests:
      description: The caller exceeded the request rate limit.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    PatientBranch:
      type: object
      additionalProperties: false
      required: [publicId, name]
      properties:
        publicId: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
        name: { type: string }
    PatientReferenceData:
      type: object
      additionalProperties: false
      required: [branches]
      properties:
        branches:
          type: array
          items: { $ref: '#/components/schemas/PatientBranch' }
    PatientReferenceResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/PatientReferenceData' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    PatientWriteRequest:
      type: object
      required: [fullName, dateOfBirth, gender]
      properties:
        fullName: { type: string, minLength: 2, maxLength: 200 }
        dateOfBirth: { type: string, format: date }
        gender: { type: string, enum: [MALE, FEMALE, OTHER] }
        nationalId: { type: [string, 'null'], maxLength: 30 }
        healthInsuranceNo: { type: [string, 'null'], maxLength: 30 }
        phone: { type: [string, 'null'], maxLength: 20 }
        email: { type: [string, 'null'], maxLength: 254 }
        addressLine: { type: [string, 'null'], maxLength: 300 }
        province: { type: [string, 'null'], maxLength: 100 }
...
    PatientDuplicateRequest:
      allOf:
        - $ref: '#/components/schemas/PatientWriteRequest'
        - type: object
          required: [branchPublicId]
          properties:
            branchPublicId: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
    CreatePatientRequest:
      allOf:
        - $ref: '#/components/schemas/PatientDuplicateRequest'
        - type: object
          properties:
            duplicateOverride: { type: boolean, default: false }
            duplicateReason: { type: string, minLength: 10, maxLength: 500 }
...
    PatientEmergencyContact:
      type: object
      additionalProperties: false
      required: [fullName, relationshipName, phone, isPrimary]
      properties:
        fullName: { type: string, minLength: 2, maxLength: 200 }
        relationshipName: { type: [string, 'null'], maxLength: 80 }
        phone: { type: string, minLength: 7, maxLength: 20, pattern: '^[0-9+(). -]+$' }
        isPrimary: { type: boolean }
...
    PatientSummary:
      type: object
      required: [publicId, code, fullName, dateOfBirth, gender, phone, status, nationalIdLast4, rowVersion]
      properties:
        publicId: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
        code: { type: string }
        fullName: { type: string }
        dateOfBirth: { type: string, format: date }
        gender: { type: string, enum: [MALE, FEMALE, OTHER] }
        phone: { type: [string, 'null'] }
        status: { type: string }
        nationalIdLast4: { type: [string, 'null'] }
        rowVersion: { type: string, pattern: '^[A-Za-z0-9+/]{11}=$' }
    PatientDetail:
      allOf:
        - $ref: '#/components/schemas/PatientSummary'
        - type: object
          required: [branch, nationalId, healthInsuranceNo, email, addressLine, province, emergencyContacts]
          properties:
            branch: { $ref: '#/components/schemas/PatientBranch' }
            nationalId: { type: [string, 'null'] }
            healthInsuranceNo: { type: [string, 'null'] }
            email: { type: [string, 'null'] }
            addressLine: { type: [string, 'null'] }
            province: { type: [string, 'null'] }
            emergencyContacts:
              type: array
              items: { $ref: '#/components/schemas/PatientEmergencyContact' }
...
    PatientDetailResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/PatientDetail' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
...
    ResponseMeta:
      type: object
      additionalProperties: true
    RequestId:
      type: string
      format: uuid
...
    PatientRelationship:
      type: string
      enum: [SELF, CHILD, SPOUSE, PARENT, GUARDIAN, OTHER]
    PatientLinkRequestStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, CANCELLED, EXPIRED]
    PatientLinkReferenceData:
      type: object
      additionalProperties: false
      required: [branches]
      properties:
        branches:
          type: array
          items:
            $ref: '#/components/schemas/BranchReference'
    PatientLinkReferenceResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          $ref: '#/components/schemas/PatientLinkReferenceData'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        requestId:
          $ref: '#/components/schemas/RequestId'
    CreatePatientLinkRequest:
      type: object
      additionalProperties: false
      required: [branchPublicId, patientCode, dateOfBirth, relationshipType]
      properties:
        branchPublicId:
          type: string
          format: uuid
        patientCode:
          type: string
          minLength: 2
          maxLength: 30
        dateOfBirth:
          type: string
          format: date
        relationshipType:
          $ref: '#/components/schemas/PatientRelationship'
        requestNote:
          type: string
          maxLength: 500
          description: Required for a relationship other than SELF.
    PatientProfileReference:
      type: object
      additionalProperties: false
      required: [publicId, code, fullName, dateOfBirth]
      properties:
        publicId:
          type: string
          format: uuid
        code:
          type: string
        fullName:
          type: string
        dateOfBirth:
          type: string
          format: date
    PatientLinkUserReference:
      type: object
      additionalProperties: false
      required: [publicId, displayName]
      properties:
        publicId:
          type: string
          format: uuid
        displayName:
          type: string
    PatientLinkVerifiedBranch:
      type: object
      additionalProperties: false
      required: [publicId, name]
      properties:
        publicId:
          type: string
          format: uuid
        name:
          type: string
    PatientAccessLink:
      type: object
      additionalProperties: false
      required:
        - publicId
        - patient
        - relationshipType
        - status
        - bookingAllowed
        - accessKind
        - linkedUser
        - verifiedBranch
        - verifiedAtUtc
        - canRevoke
        - rowVersion
      properties:
        publicId:
          type: string
          format: uuid
        patient:
          $ref: '#/components/schemas/PatientProfileReference'
        relationshipType:
          $ref: '#/components/schemas/PatientRelationship'
        status:
          type: string
          const: ACTIVE
        bookingAllowed:
          type: boolean
        accessKind:
          type: string
          enum: [OWN, DELEGATED]
        linkedUser:
          $ref: '#/components/schemas/PatientLinkUserReference'
        verifiedBranch:
          oneOf:
            - $ref: '#/components/schemas/PatientLinkVerifiedBranch'
            - type: 'null'
        verifiedAtUtc:
          type: [string, 'null']
          format: date-time
        canRevoke:
          type: boolean
        rowVersion:
          type: string
          pattern: '^[A-Za-z0-9+/]{11}=$'
    PatientLinkRequest:
      type: object
      additionalProperties: false
      required:
        - publicId
        - branch
        - patientReference
        - relationshipType
        - requestNote
        - status
        - decisionReason
        - createdAtUtc
        - expiresAtUtc
        - decidedAtUtc
        - patient
        - rowVersion
      properties:
        publicId:
          type: string
          format: uuid
        branch:
          $ref: '#/components/schemas/BranchReference'
        patientReference:
          type: string
        relationshipType:
          $ref: '#/components/schemas/PatientRelationship'
        requestNote:
          type: [string, 'null']
        status:
          $ref: '#/components/schemas/PatientLinkRequestStatus'
        decisionReason:
          type: [string, 'null']
        createdAtUtc:
          type: string
          format: date-time
        expiresAtUtc:
          type: string
          format: date-time
        decidedAtUtc:
          type: [string, 'null']
          format: date-time
        patient:
          oneOf:
            - type: object
              additionalProperties: false
              required: [publicId, code, fullName]
              properties:
                publicId:
                  type: string
                  format: uuid
                code:
                  type: string
                fullName:
                  type: string
            - type: 'null'
        rowVersion:
          type: string
          pattern: '^[A-Za-z0-9+/]{11}=$'
...
    PatientAccessData:
      type: object
      additionalProperties: false
      required: [links, requests]
      properties:
        links:
          type: array
          items:
            $ref: '#/components/schemas/PatientAccessLink'
        requests:
          type: array
          items:
            $ref: '#/components/schemas/PatientLinkRequest'
    PatientAccessResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          $ref: '#/components/schemas/PatientAccessData'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        requestId:
          $ref: '#/components/schemas/RequestId'
    PatientLinkRequestAcceptedData:
      type: object
      additionalProperties: false
      required: [requestId, status]
      properties:
        requestId:
          type: string
          format: uuid
        status:
          type: string
          const: PENDING
    PatientLinkRequestAcceptedResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          $ref: '#/components/schemas/PatientLinkRequestAcceptedData'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        requestId:
          $ref: '#/components/schemas/RequestId'
    PatientLinkReasonRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 3
          maxLength: 500
    PatientLinkCancelledData:
      type: object
      additionalProperties: false
      required: [cancelled]
      properties:
        cancelled:
          type: boolean
          const: true
    PatientLinkCancelledResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          $ref: '#/components/schemas/PatientLinkCancelledData'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        requestId:
          $ref: '#/components/schemas/RequestId'
    PatientLinkRevokedData:
      type: object
      additionalProperties: false
      required: [revoked]
      properties:
        revoked:
          type: boolean
          const: true
    PatientLinkRevokedResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          $ref: '#/components/schemas/PatientLinkRevokedData'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        requestId:
          $ref: '#/components/schemas/RequestId'
...
    BranchReference:
      type: object
      additionalProperties: false
      required: [publicId, code, name]
      properties:
        publicId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
...
    ReceptionPatient:
      type: object
      additionalProperties: false
      required: [publicId, code, fullName, dateOfBirth, gender, phone]
      properties:
        publicId: { type: string, format: uuid }
        code: { type: string }
        fullName: { type: string }
        dateOfBirth: { type: string, format: date }
        gender: { type: string, enum: [MALE, FEMALE, OTHER] }
        phone: { type: [string, 'null'] }
    ReceptionPatientListResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/ReceptionPatient' } }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
...
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required: [error, requestId]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
        requestId:
          $ref: '#/components/schemas/RequestId'
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true